split code up in multiple files for easier management, update requirements.txt to use specific versions, fixes #2
